# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClassifyInput {
  /** The physics question or query to classify */
  query: string;
  /**
   * Base64-encoded image providing visual context
   * @nullable
   */
  imageContext?: string | null;
}

export interface QueryIntent {
  /** The classified intent category identifier */
  query_intent: string;
  /** The core physics concept involved */
  physics_concept: string;
  /** The best visual type to use */
  visual_type: string;
  /** List of recommended framework identifiers */
  recommended_frameworks: string[];
  requires_animation: boolean;
  requires_numeric_data: boolean;
}

export interface GenerateQuestionInput {
  /** The category ID to generate a question for */
  categoryId: string;
  /** Optional example question to base the new question on */
  exampleQuestion?: string;
}

export interface GeneratedQuestion {
  question: string;
}

export interface VisualizeInput {
  query: string;
  intent: QueryIntent;
  /**
   * Base64-encoded image for context (graph/diagram the user is asking about)
   * @nullable
   */
  imageContext?: string | null;
}

export interface MatplotlibOutput {
  /**
   * Base64-encoded PNG image
   * @nullable
   */
  imageBase64: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  renderTimeMs?: number | null;
}

export type RechartsOutputDataItem = { [key: string]: unknown };

export interface RechartsOutput {
  /**
   * e.g. line, bar, area
   * @nullable
   */
  chartType?: string | null;
  /**
   * Array of data points
   * @nullable
   */
  data?: RechartsOutputDataItem[] | null;
  /** @nullable */
  xKey?: string | null;
  /** @nullable */
  yKeys?: string[] | null;
  /** @nullable */
  colors?: string[] | null;
  /** @nullable */
  xLabel?: string | null;
  /** @nullable */
  yLabel?: string | null;
  /** @nullable */
  title?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}

export interface D3Output {
  /**
   * SVG markup string
   * @nullable
   */
  svgContent?: string | null;
  /** @nullable */
  diagramType?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}

export interface ManimOutput {
  /**
   * Manim Python script
   * @nullable
   */
  script?: string | null;
  /** @nullable */
  sceneClass?: string | null;
  /**
   * CSS-animated SVG preview of the physics motion
   * @nullable
   */
  animatedSvg?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}

export interface GoogleImageOutput {
  /**
   * Base64-encoded image
   * @nullable
   */
  imageBase64?: string | null;
  /** @nullable */
  mimeType?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}