# lib/api-zod/src/generated/types/d3Output.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface D3Output {
  /**
   * SVG markup string
   * @nullable
   */
  svgContent?: string | null;
  /** @nullable */
  diagramType?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}