# lib/api-zod/src/generated/types/generateQuestionInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateQuestionInput {
  /** The category ID to generate a question for */
  categoryId: string;
  /** Optional example question to base the new question on */
  exampleQuestion?: string;
}