# lib/api-zod/src/generated/types/googleImageOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoogleImageOutput {
  /**
   * Base64-encoded image
   * @nullable
   */
  imageBase64?: string | null;
  /** @nullable */
  mimeType?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}