# lib/api-zod/src/generated/types/manimOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ManimOutput {
  /**
   * Manim Python script
   * @nullable
   */
  script?: string | null;
  /** @nullable */
  sceneClass?: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
}