# lib/api-zod/src/generated/types/matplotlibOutput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatplotlibOutput {
  /**
   * Base64-encoded PNG image
   * @nullable
   */
  imageBase64: string | null;
  suitable: boolean;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  renderTimeMs?: number | null;
}