# lib/api-zod/src/generated/types/visualizeInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QueryIntent } from "./queryIntent";

export interface VisualizeInput {
  query: string;
  intent: QueryIntent;
}